django/pyproject.toml
[tool.setuptools]
packages = ["core", "base", "crawler"]

[project]
name = "Subscription_Service"
version = "0.1.0"
description = "A django project that uses kafka to send data to email sends service and also uses celery to do tasks"
authors = [{name = "Nima Hekmati", email = "hekmatinima@gmail.com"}]
requires-python = "~=3.12"
dependencies = [
    "Django==5.1",
    "djangorestframework==3.15.2",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "celery==5.4.0",
    "python-decouple==3.8",
    "psycopg2-binary==2.9.10",
    "kafka-python==2.0.4",
    "redis==5.2.1",
    "gunicorn>=23.0.0",
]



[dependency-groups]
dev = [
    "django-stubs>=5.1.3",
    "go-task-bin>=3.42.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.5.2",
    "ruff>=0.9.10",
    "uv>=0.6.3",
]



[tool.uv]
index-url = "https://pypi.org/simple/"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "my_django_project.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
